add qtash to manage auctions
